+++ include/ata.h
@@ -8,13 +8,13 @@
 #include <variant>
 #include <vector>
 
-#define ATA_VERSION "1.6.1"
+#define ATA_VERSION "1.6.2"
 
 namespace ata {
 
 inline constexpr uint32_t VERSION_MAJOR = 1;
 inline constexpr uint32_t VERSION_MINOR = 6;
-inline constexpr uint32_t VERSION_REVISION = 1;
+inline constexpr uint32_t VERSION_REVISION = 2;
 
 inline constexpr std::string_view version() noexcept {
   return ATA_VERSION;